Update MonsterAssessment_Python.py
Added easygui to the function, allowing us to return to a home screen and given the options to pick from various branches. I made it so you can view what the code actually is in pop up form, giving us more freedom on what we pick to do with our monster cards. Beginning to work on the 'adding new card' function.